Q4 parallel done. Need to add result in parallel and output time of computation

+++ ECSE420A3/src/ca/mcgill/ecse420/a3/q4/MatrixHelper.java
@@ -5,7 +5,7 @@ import java.util.Random;
 
 public class MatrixHelper {
 
-	public static void printMatrix(double[][] matrix) {
+	public static void printMatrix(int[][] matrix) {
 		for (int j = 0; j < matrix.length; j++) {
 			for (int i = 0; i < matrix[j].length; i++) {
 				System.out.print(matrix[i][j] + " ");
@@ -15,7 +15,7 @@ public class MatrixHelper {
 		System.out.println();
 	}
 
-	public static void printVector(double[] vector) {
+	public static void printVector(int[] vector) {
 		System.out.print(vector[0]);
 		for (int i = 1; i < vector.length; i++) {
 			System.out.print(" , " + vector[i]);
@@ -23,24 +23,24 @@ public class MatrixHelper {
 		System.out.println();
 	}
 
-	public static double[][] randomMatrixGenerator(int n, int m) {
-		double[][] randomMatrix = new double[n][m];
+	public static int[][] randomMatrixGenerator(int n, int m) {
+		int[][] randomMatrix = new int[n][m];
 		for (int i = 0; i < n; i++) {
 			randomMatrix[i] = randomVectorGenerator(m);
 		}
 		return randomMatrix;
 	}
 
-	public static double[] randomVectorGenerator(int vectorSize) {
+	public static int[] randomVectorGenerator(int vectorSize) {
 		Random rand = new Random();
-		double[] randomVector = new double[vectorSize];
+		int[] randomVector = new int[vectorSize];
 		for (int i = 0; i < vectorSize; i++) {
-			randomVector[i] = rand.nextDouble() + 1;
+			randomVector[i] = rand.nextInt() + 1;
 		}
 		return randomVector;
 	}
 
-	public static boolean validateVectorResult(double[] a, double[] b) {
+	public static boolean validateVectorResult(int[] a, int[] b) {
 		return Arrays.equals(a, b);
 	}
